requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pymysql
python-dotenv
httpx
pillow
